+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## 0.1.2 - 2016-01-11
 ### Added
 - SdCardStream
+### Tested
+- compilation with Arduino UNO, DUE
 
 ## 0.1.1 - 2016-01-11
 ### Fixed